test(integration): share isolated server harness
